+++ src/application/App.java
@@ -4,7 +4,6 @@ import java.io.IOException;
 import java.net.UnknownHostException;
 import java.sql.SQLException;
 
-import application.Database.Role;
 import application.controllers.Controller;
 import javafx.application.Application;
 import javafx.fxml.FXMLLoader;
@@ -16,7 +15,7 @@ public class App extends Application {
 
 	@Override
 	public void start(Stage primaryStage) throws IOException, Exception {
-		Database.connectAs(Role.NEUTRAL);
+		Database.connect();
 		loadPage("TestView", primaryStage);
 		primaryStage.setWidth(1000);
 		primaryStage.setHeight(600);
@@ -60,10 +59,5 @@ public class App extends Application {
 }
 
 // TODO: Avoid using String.format() when querying the database
-// TODO: Make an Editable interface for ValueField, ValueOption, ValueLabel, and ValueRow
 // TODO: When edit is clicked, a RadioButton should appear next to each row, and the user should be able to select multiple rows to delete
 // TODO: Test insertng and deleting visits as a patient and doctor
-// TODO: Rows such as visits should not be deletable with a button. Instead, they should be deletable by clicking on the row and pressing the delete button from the popup
-// TODO: Information rows such as surgeries or allergies should be deletable with a delete button
-// TODO: Create a ValueRow instead of ValueLabel and it shall consist of Labels
-// TODO: Change Value Interface to Updatable
